This version has two fixes:
-> Cubed AutoNormals no longer crash with voxels bigger than 250 voxels wide at any axis, as reported by Tony
http://www.ppmsite.com/forum/viewtopic.php?p=359842#359842

-> Mesh Smooth operation at the 3D Modelizer window works much faster and using less RAM.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@221 51f8c993-4936-0410-a6d7-a9c99ad3dd25

+++ vxlseiii14x/Voxel_Tools.pas
@@ -132,7 +132,7 @@ var
    Res : TVector3f;
    Sum : TVector3f;
    {Alpha,}x,y,z : integer;
-   xx,yy,zz : byte;
+   xx,yy,zz : integer;
    MyPoint : TVector3i;
    MinPoint,MaxPoint : TVector3i;
    MidPointMinusRange : integer;
